Header data from list container/object is now saved on the entity

+++ src/main/java/nl/tweeenveertig/openstack/command/account/ContainerListElement.java
@@ -4,7 +4,7 @@ public class ContainerListElement {
 
     public String name;
 
-    public long count;
+    public int count;
 
     public long bytes;
